# DistanceOp: Optimize some implementation details