- Add short-circuits to Extracter classes
- Avoid heap allocation of GeometryLocation

main:
```
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
BM_PointPointDistance       2.15 ms         2.15 ms          315
BM_PointLineDistance        60.4 ms         60.4 ms           11
BM_LineLineDistance         41.5 ms         41.5 ms           17
```

this PR:
```
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
BM_PointPointDistance       2.10 ms         2.10 ms          329
BM_PointLineDistance        39.5 ms         39.5 ms           18
BM_LineLineDistance         35.0 ms         35.0 ms           20
```